# listListenerMatching optimization