@andrewluetgers 2015-04-10
@mikaelbr here is a Map implementation with deepSet and deepGet demonstrating a data-structure that could be used in place of this current implementation using basic JS objects with string keys. It also shows how the data-structure would store the callbacks and how you could add and retrieve them.
https://jsfiddle.net/aq93v4g7/3/