@Dashed 2015-04-10
Looks like something similar to `deepSet`/`deepGet` is being incubated in lodash as `_.get` and `_.set`: https://github.com/lodash/lodash/commit/2edcc893034936cccb94264783453b3d6bceff45

Might be good to use that in the future, as alternative implementations, when it gets released.