@andrewluetgers 2015-04-10
@Dashed not exactly, this is a very specific use case and not used anywhere else in the codebase. I contributed to [lodash-deep](https://github.com/marklagendijk/lodash-deep) early on and am pretty familiar with the issues involved. Fundamentally you can cut a lot of corners in this case as it is an optimization and not a general use utility there should be no problem with that.