@Dashed 2015-04-10
Fair points. But there's no harm in outsourcing these utilities to eventual modules:
https://npm.im/lodash.get
https://npm.im/lodash.set

I suggest this because those API methods may have the necessary optimizations you don't need to worry about, nor maintain at a higher-level library like immstruct.