@andrewluetgers 2015-04-10
This is the thing, there is not going to be a deepGet, or deepSet implementation in lodash that works with a Map data-structure (takes any value for keys, uses Map.get, Map.set), it will work with native js objects and arrays (strings for keys, native object access and mutation syntax). Similarly I also could write this optimization with the Map implementation in immutable.js that would obviate the use of my custom Map but that would be overkill for a performance optimization I suspect it would use more memory and be slower. Of course only testing both will tell you that but at either rate you need some code to take a keyPath into a Map and use it to set and get values to make it all work.

https://gist.github.com/andrewluetgers/7f0f043ea2224c315ad0#file-map-js-L37-L67