@andrewluetgers 2015-04-10
when lodash releases a deep get, deep set I'll certainly look at it for some possible improvements but if I want to use their code I'll have to fork it to support working with a tree of Maps. There is already a [lodash-deep](https://github.com/marklagendijk/lodash-deep) which I have actually contributed to. I could have also used that but again it would be a fork and there is much complexity there because it is intended as a general use utility.