@andrewluetgers 2015-04-11
@mikaelbr I have merged in the Map changes so now this change supports arbitrary types for keys. I think this still needs more tests but would like your feedback. Performance-wise there is no discernible impact to switching to the new Map structure. http://plnkr.co/edit/TRh18pmklgOF8cbunYbj?p=preview