@mikaelbr 2015-04-11
I was experimenting with maps yesterday my self, but since we have access to Immutable.js I thought it made more sense to use Immutabe.Map. I had some kinks I needed to work out, though. But maybe it makes sense to use Immutable.Map? Also, while I was looking into the source code yesterday, I fpund thee code a bit hard to follow and I'm not the biggest fan of mutation and side-effect in the helper functions (match listeners etc).