@andrewluetgers 2015-04-11
I'm not sure why you need immutability now to back the storage of callbacks, was that a concern before when you were using arrays? I think the two (needing a Map vs. needing immutability) are getting conflated.