@mikaelbr 2015-04-11
I don't say we need immutability, but we already have a Map implementation available - and probably don't need [a custom implementation](https://github.com/andrewluetgers/immstruct/commit/2fab8fc4fe26ea306559347dd002a45607d81cdc#diff-2b4ca49d4bb0a774c4d4c1672d7aa781R10)