@Dashed 2015-04-11
I glanced over the rework using the Map implementation. I noticed that appending `__listeners__` to path doesn't actually avoid collisions:

```js
var _pathListeners = Map();
getListenerNs(_pathListeners, ['foo']).push(1);
getListenerNs(_pathListeners, ['foo', '__listeners__']); // => error
```

A compromise may be to make `__listeners__` configurable via Structure options.
Or, if you don't want an official API for it, make it a property `structure._listenersKey = '__listeners__'` for advanced users.