@andrewluetgers 2015-04-11
@Dashed yes, it is not perfect, I think it would be good to throw friendly error message if we see a key that matches the listeners key and make it configurable.