@andrewluetgers 2015-04-11
@Dashed I just realized now that it uses a Map the listeners key can be a unique sentinel value ala
  var listenersSentinel = {};
this would fully prevent any collisions :-)