@Dashed 2015-04-11
@andrewluetgers You should probably put this map implementation as an npm module.