@mikaelbr 2015-04-11
@andrewluetgers I added a [PR](https://github.com/andrewluetgers/immstruct/pull/1) for replacing the Map with a already available Map, and re-wrote some functions to show better intent. The thing I thought was making it harder to follow was that _pathListeners was changed as a side-effect inside `getListenerNs`. I'd much rather have pure functions - which only change the system by its return value.