@andrewluetgers 2015-04-12
@mikaelbr @Dashed are either of you already or wanting to work on implementing the fix for cursor specific callbacks using cursor sentinel values, if not I will.