@andrewluetgers 2015-04-17
@Dashed @mikaelbr please take a look at the latest changes for unobserving, THANKS!