@mikaelbr 2015-04-28
@andrewluetgers Looks good to me. I see you extended the `unobserveAll` API to include "unobserve all but this". Must remember to document that.

I feel comfortable merging this as you've added the test of `unobserve` not leaking to other references, and all other tests pass. Thanks for your great work!